+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a custom event handler to the session timeout modal JS to allow applications to trigger keepalive requests. (AUTHLIB-171)
+- Added the ability to configure the debounce timeout of the `octri.keepalive` event. (AUTHLIB-171)
 
 ## [4.0.0] - 2026-03-17